pyproject.toml
[project]
name = "srchive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.2.0",
    "dnspython>=2.7.0",
    "pinecone>=7.3.0",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.46.1",
    "tqdm>=4.67.1",
]
